moving / renaming packages, fixing stuff
