+++ CommandLine/src/main/java/command/Main.java
@@ -1,12 +1,7 @@
 package command;
 
-import fsobject.Directory;
-
 public class Main {
 	public static void main(String[] args) {
-
-		new Directory();
-
 		System.out.println("hello world!");
 	}
 }
