+++ CommandLine/src/main/java/file_system_adapter/FakeFileSystemAdapter.java
@@ -1,11 +1,6 @@
 package file_system_adapter;
 
-import fileSystemObject.FakeDirectory;
-import fileSystemObject.FakeFSO;
-import fsobject.Directory;
-
-import java.io.File;
-import java.util.ArrayList;
+import file_system_adapter.fakeFSO.FakeDirectory;
 
 public class FakeFileSystemAdapter implements FileSystemAdapter {
 	public static final String DIR_SEPERATOR = "/";
